refactor: 토너먼트 라우팅을 [id] path param으로 변경
